requirements.txt
Flask==3.0.0
openai-whisper==20231117
pytest==7.4.4
redis==5.0.1
rq==1.16.2
python-dotenv==1.0.0
gunicorn==21.2.0
sentry-sdk==1.39.1
blinker==1.7.0
Werkzeug==3.0.1
Unidecode==1.3.8
requests==2.31.0
requests-mock==1.11.0
autopep8==2.0.4
flasgger==0.9.7.1
pika==1.3.2